+++ app/TypiCMS/Modules/Places/Presenters/ModulePresenter.php
@@ -5,13 +5,4 @@ use TypiCMS\Presenters\Presenter;
 
 class ModulePresenter extends Presenter
 {
-    /**
-     * Get title
-     * 
-     * @return string
-     */
-    public function title()
-    {
-        return $this->entity->title;
-    }
 }