+++ app/TypiCMS/Modules/Places/Views/places/admin/_form.blade.php
@@ -12,52 +12,66 @@
 
 @include('admin._buttons-form')
 
-<div class="row">
+{{ Form::hidden('id') }}
 
-    {{ Form::hidden('id'); }}
+@include('admin._image-fieldset', ['field' => 'image'])
 
-    <div class="col-sm-6">
+<ul class="nav nav-tabs">
+    <li class="active">
+        <a href="#tab-main" data-target="#tab-main" data-toggle="tab">@lang('global.Content')</a>
+    </li>
+    <li>
+        <a href="#tab-info" data-target="#tab-info" data-toggle="tab">@lang('global.Info')</a>
+    </li>
+</ul>
 
-        <div class="form-group @if($errors->has('title'))has-error @endif">
-            {{ Form::label('title', trans('validation.attributes.title')) }}
-            {{ Form::text('title', null, array('class' => 'form-control')) }}
-            {{ $errors->first('title', '<p class="help-block">:message</p>') }}
-        </div>
-        <div class="form-group @if($errors->has('slug'))has-error @endif">
-            {{ Form::label('slug', trans('validation.attributes.slug'), array('class' => 'control-label')) }}
-            <div class="input-group">
-                {{ Form::text('slug', null, array('class' => 'form-control')) }}
-                <span class="input-group-btn">
-                    <button class="btn btn-default btn-slug @if($errors->has('slug'))btn-danger @endif" type="button">@lang('validation.attributes.generate')</button>
-                </span>
-            </div>
-            {{ $errors->first('slug', '<p class="help-block">:message</p>') }}
-        </div>
+<div class="tab-content">
 
-        <div class="row">
+    {{-- Main tab --}}
+    <div class="tab-pane fade in active" id="tab-main">
 
-            @foreach ($locales as $lang)
+        @include('admin._tabs-lang-form', ['target' => 'content'])
 
-            <fieldset class="col-sm-6" id="{{ $lang }}">
-                <legend>{{ trans('global.languages.'.$lang) }}</legend>
+        <div class="tab-content">
+
+        @foreach ($locales as $lang)
+
+            <div class="tab-pane fade @if($locale == $lang)in active @endif" id="content-{{ $lang }}">
+                <div class="row">
+                    <div class="col-md-6 form-group">
+                        {{ Form::label($lang.'[title]', trans('validation.attributes.title')) }}
+                        {{ Form::text($lang.'[title]', $model->translate($lang)->title, array('class' => 'form-control')) }}
+                    </div>
+                    <div class="col-md-6 form-group @if($errors->has($lang.'.slug'))has-error @endif">
+                        {{ Form::label($lang.'[slug]', trans('validation.attributes.slug'), array('class' => 'control-label')) }}
+                        <div class="input-group">
+                            {{ Form::text($lang.'[slug]', $model->translate($lang)->slug, array('class' => 'form-control')) }}
+                            <span class="input-group-btn">
+                                <button class="btn btn-default btn-slug @if($errors->has($lang.'.slug'))btn-danger @endif" type="button">@lang('validation.attributes.generate')</button>
+                            </span>
+                        </div>
+                        {{ $errors->first($lang.'.slug', '<p class="help-block">:message</p>') }}
+                    </div>
+                </div>
                 <div class="checkbox">
                     <label>
                         {{ Form::checkbox($lang.'[status]', 1, $model->translate($lang)->status) }} @lang('validation.attributes.online')
                     </label>
                 </div>
                 <div class="form-group">
-                    {{ Form::label($lang.'[info]', trans('validation.attributes.info')) }}
-                    {{ Form::textarea($lang.'[info]', $model->translate($lang)->info, array('class' => 'form-control', 'rows' => 5)) }}
+                    {{ Form::label($lang.'[body]', trans('validation.attributes.body')) }}
+                    {{ Form::textarea($lang.'[body]', $model->translate($lang)->body, array('class' => 'editor form-control')) }}
                 </div>
-            </fieldset>
+            </div>
 
-            @endforeach
+        @endforeach
 
         </div>
 
     </div>
 
-    <div class="col-sm-6">
+    {{-- Info tab --}}
+    <div class="tab-pane fade in" id="tab-info">
 
         <div class="form-group">
             {{ Form::label('address', trans('validation.attributes.address'), array('class' => 'control-label')) }}
@@ -98,8 +112,6 @@
                 {{ Form::text('longitude', null, array('class' => 'form-control')) }}
             </div>
         </div>
-
-        @include('admin._image-fieldset', ['field' => 'image'])
 
     </div>