+++ app/TypiCMS/Modules/Places/Views/places/public/show.blade.php
@@ -1,36 +1,9 @@
-<!doctype html>
-<html lang="fr">
+@section('js')
+    {{ HTML::script(asset('//maps.googleapis.com/maps/api/js?sensor=false&amp;language='.Config::get('app.locale'))) }}
+    {{ HTML::script(asset('js/public/gmaps.js')) }}
+@stop
 
-<head>
-
-    <meta charset="utf-8">
-    <title>{{ $model->title }}</title>
-
-    {{ HTML::style(asset('css/gmaps.css')) }}
-    {{ HTML::style(asset('css/public.css')) }}
-    {{ HTML::style(asset('vendor/fancybox/source/jquery.fancybox.css')) }}
-
-    {{ HTML::script(asset('vendor/jquery-legacy/jquery.js')) }}
-    {{ HTML::script(asset('vendor/bootstrap/js/dropdown.js')) }}
-    {{ HTML::script(asset('vendor/fancybox/source/jquery.fancybox.pack.js')) }}
-    {{ HTML::script(asset('//maps.googleapis.com/maps/api/js?sensor=false&amp;language=fr')) }}
-    {{ HTML::script(asset('js/gmaps.js')) }}
-
-    @if(Config::get('typicms.typekitCode'))
-    <script type="text/javascript" src="//use.typekit.net/{{ Config::get('typicms.typekitCode') }}.js"></script>
-    <script type="text/javascript">try {Typekit.load();} catch (e) {}</script>
-    @endif
-
-    <!--[if lt IE 9]>
-    <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-    <script src="https://oss.maxcdn.com/libs/respond.js/1.3.0/respond.min.js"></script>
-    <![endif]-->
-
-    {{ HTML::script(asset('js/public.js')) }}
-
-</head>
-
-<body style="padding:0">
+@section('main')
 
     <div class="row">
         <div class="col-sm-4">
@@ -67,6 +40,4 @@
         </div>
     </div>
 
-</body>
-
-</html>
+@stop